Modularize notification systems
